fix: fix bug for otp & qr confirmation strategies

+++ src/main/java/com/podzilla/courier/dtos/delivery_tasks/SubmitCourierRatingResponseDto.java
@@ -1,4 +1,6 @@
 package com.podzilla.courier.dtos.delivery_tasks;
 
-public record SubmitCourierRatingResponseDto(String id, String orderId, String courierId, Double courierRating) {
+import java.math.BigDecimal;
+
+public record SubmitCourierRatingResponseDto(String id, String orderId, String courierId, BigDecimal courierRating) {
 }
